refactor(ensure): port index(.js) and its tests to ts
Also updated globby to latest, as it has typescript definition files.
