+++ @commitlint/ensure/package.json
@@ -65,7 +65,7 @@
     "babel-register": "6.26.0",
     "concurrently": "3.5.1",
     "cross-env": "5.1.1",
-    "globby": "8.0.1",
+    "globby": "^9.2.0",
     "jest": "^24.8.0",
     "ts-jest": "^24.0.2",
     "typescript": "^3.4.5"
